Fix web server so start up doesn't emit meaningless error message
When starting up with the default configuration the web server emits a meaningless error message.  This is because the code
to process the repeatHeader option doesn't first check if option has been set and, by default, it isn't.

The following patch adds a test around the code to prevent the error message.  It also fixes the misspelling as a side-effect
of making the code DRYer.

The patch also resolves star imports, adds @Override annotation and removes an unnecessary String cast.

RB: http://rb.dcache.org/r/593/
Acked-by: Gerd