core: Adds retryOnNoRouteToCell property to CellStub
This property controls what happens when sending a message fails with a
NoRouteToCellException. When the property is set to true, the operation
will be retried until the timeout has been reached. If false, the
operation will fail right away.

PnfsHandler used to have the former behavior, as any communication
failure with PnfsManager is likely to be temporary. However since
PnfsHandler was updated to use CellStub, this functionality was lost.
With this patch the functionality is reintroduced.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/589/