dir: Use new list framework for directory listing
Updates DirectoryLookUpPool to use new the new PnfsManager based 
directory listing. As /pnfs/fs is not a valid path when talking to the 
PnfsManager, I had to change the root path to /. This only affects the 
administrative ls command, so this should not be a big problem. 

Paul and Patrick suggested to completely remove the option to alter the 
root path, as it only affected the administrative command anyway. Thus 
the directoryPoolPnfsBase parameter has been removed.

Target: trunk
Patch: http://rb.dcache.org/r/595/
Acked-by: Paul Millar <paul.millar@desy.de>