Add more debugging information to detect problems in PNFS mount detection
The following patch adds log4j debugging output for the PNFS automatic mount-point detection and for the steps required to accept a PNFS mount-point.

If a mount-point is rejected by BasicNamespaceProvider there is little reason given as to why.  This patch addresses this problem by emitting whether, for each candidate mount-point, whether it is accepted or not.  It also emits information if the VirtualMountPoint cannot be created.

Acked-by: Gerd
ReviewBoard: http://rb.dcache.org/r/602/