pnfsmanager+ftpdoor: Permission check in name space provider
This patch adds PermissionHandlerNameSpaceProvider to perform permission
checking in the PnfsManager. The classic PermissionHandler based
permission check in the FTP door can be disabled by using a
GrantAllPermissionHandler. PnfsHandler is extended to allow setting a
subject and the FTP door is extended to set that subject.

Whether to use PnfsManager based permission checking or not is
controlled by whether a non-root subject is set in the request. Whether
to do permission checking in the FTP door is controlled by whether a
permission handler is configured or not. If a permission handler is not
configured in the door, then the door sets a non-root subject in
PnfsManager requests, thus enforcing permission checking in PnfsManager.

Target: trunk
Acked-by Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/475/