doors: Add Subject field to DirectoryListSource
Not all doors have a PnfsHandler per user session. In particular the SRM
door uses a single cell instance to process requests from different
users. Therefore relying on the setSubject call of PnfsHandler does work
for the SRM door.

This patch adds a subject parameter to the DirectoryListSource methods.
This is so we can start using it in the SRM door.

The patch also adds a copy constructor to PnfsHandler. This is used for
creating user specific "copies" of PnfsHandler. I realize this is not
particularly clean, but I see this as a temporary solution until we
replace PnfsHandler with Tigran's new client stub.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/611/