Add StateExhibitor interface
The following patch adds the StateExhibitor interface.  Classes that implement this interface
allow the visitation of dCache state using the visitor pattern.

With the introduction of this interface, the constructor of classes that (directly or indirectly)
must access dCache state have been updated to include an explicit StateExhibitor as a parameter.
In addition to opening up the possibility of unit-testing these classes, this also may be viewed
as the first step towards removing the singleton nature of the State class.

Acked-by: Tigran
RB: http://rb.dcache.org/r/619/