Add new classes to facilitate unit testing of info service
TestStateExhibitor is a self-contained, complete implementation of the StateExhibitor interface.
Metrics may be easily added and the resulting object can be visited by any class implementing the
StateVisitor interface.  The TestStateExhibitorTests class provides unit tests that this class
behaves as expected.

The MalleableStateTransition provides a convenient class for testing behaviour of the info service
when one or more metrics alter (added, deleted, updated).  This class facilitates the unit testing
of those classes designed to maintain "secondary" (or "derived") data.  The
MalleableStateTransitionTests class provides unit tests for MalleableStateTransition, although it
depends on TestStateExhibitor to achieve them.