srm: Use DirectoryListSource to obtain directory listing
This patch updates the SRM Storage class to use a DirectoryListSource to
obtain directory listing.

There are three list methods and all three are updated. Notice that one
of these is written specifically to be used as part of a recursive rmdir
operation; therefore it performs permission checking as a side effect.
This is not particularly beautiful, but it is how the interface was
designed.

Notice that the SRM door still requires the mounted file system: The
generic SRM code in some places assumes a mounted file system and this
patch does not change that.

Also notice that this patch is unlikely to provide any of the speed
gains we have seen in the FTP door. To achieve those gains we need to
change the interface of AbstractStorageElement to provide a richer
directory listing than just the names of the entries: The benefit of the
new PnfsManager based listing is that we can query most of the
attributes of a file through the list operation.

I propose we fix both of these issues after this patch.

As DirectoryListHandler requires that we route reply messages to it, the
patch refactor messageArrived of the Storage class.

In contrast to the old code, the new code uses the proper subclasses of
SRMException for the different error cases.

Permission handling for list operations are delegated to PnfsManager.
Thus they are now subject to ACL checks.

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/616/