Fix HSM restore of zero length files
Previous versions of dCache would prevent the HSM script from running for zero length files.  This explicit behaviour was dropped 
in dCache 1.9.1.  It was anticipated that HSM scripts would filter out these zero length files if necessary (or desirable).

It turns out that the pool kept a remnant of this treating zero-length files as special: it would flag any files with (apparently) 
zero-length an error and would throw as CacheException (rc=666) with message "Couldn't get file size of [ID]".

This patch fixes this by removing this erroneous check.