Add StateObservatory interface
The following patch add the Interface StateObservatory.  This Interface encapsulates the
concept of a container that allows the registration of StateWatcher objects.  It also
provides a mechanism to allow them to being informed about changes to dCache's state and
collect any new metrics that they might produce.

Acked-by: Gerd
RB: http://rb.dcache.org/r/639