dcap: Add PnfsManager based permission checking
This patch makes the same change to the DCAP door as r12151 made to the
FTP door: When the permission handler configuration is set to the empty
string, then permission checking in the DCAP door is replaced by
permission checking in PnfsManager.

To ensure that we perform the same checks we did in the DCAP door, I
added permission checking for setFileAttributes and setFileMetaData in
PermissionHandlerNamespaceProvider. I also updated the semantics of
canSetAttributes so that MODE, UID and GID can only be changed by the
owner and all other attributes can be changed if the caller has write
permissions to the file (rather than the parent directory).

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/641