Stage configuration in PoolManager
This patch adds optional stage permission checking to PoolManager. This
has the benefit that stage configuration does not need to be maintained
on all doors.

If stage permission checking is configured in doors, then the
PoolManager will not repeat the check. We may consider to alter this
such that both the door and the pool manager perform permission
checking, if configured. The code would be more natural if we made this
change.

This code was originally developed by Irina and submitted in
http://rb.dcache.org/r/564/. I refactored the code, fixed a couple of
bugs and also added support for the GSI DCap door and the pin manager.

Stage configuration related part of batch files has been moved to
config/stage.fragment.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/604/