PnfsManager: Add error reporting and overwrite flag to renameEntry
This will make it simpler to implement renaming in doors.

Besides reporting why a rename operation has failed, the patch adds an
option to the rename message to suppress overwriting. Notice that the
semantics is not strict: With both PNFS and Chimera there is a potential
race condition if the destination name is created between the check and
the actual rename operation. However putting the check into the rename
operation rather than letting the door check the existance of the target
name reduces the window of the race condition.

The PermissionHandlerNameSpaceProvider has been extended to check delete
permissions on the target file.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/628