This patch provides three interrelated activities.
First, it introduces the StateUpdateManager interface.  This interface abstracts
StateMaintainer: introducing the general concept of a component that other components
can entrusted fresh StateUpdate object with.  It is expected that the hand-over of a
StateUpdate to the StateUpdateManager is quick (non-blocking).

Second, the patch introduces a new StateMaintainer that is much simpler.  Also included
is extensive unit-testing.

Finally, the patch removes the singleton nature of State.  To achieve this, we must pass
a StateUpdateManager object to those elements that may wish to update dCache state.  Although
this may appear clumsy, it has the benefit of allowing easier and finer-grain unit testing
of those classes.

RB: http://rb.dcache.org/r/640/
Acked-by: Gerd