srm: Refactor srm move to rely on PnfsManager for permission checking
Due to the improved error handling and permission checking in
PnfsManager, the move code could be simplified.

I extended the PnfsHandler and rename messages such that the source file
can be specified by name rather than PNFS ID. This saves a round trip to
the PnfsManager.

I had to update the PermissionHandlerNameSpaceProvider to translate a
FileNotFoundCacheException to NotDirCacheException in case the
destination directory does not exist (this is such that the caller can
distinguish it from a missing source file).

Testing Done:

Tried various srmmv combinations to cover both success and failure 
cases.

Target: trunk
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/656/