Remove singleton nature of InfoProvider class, add two interfaces to simplify interactions.
This patch is similar to the previous patch: its aim is to remove a singleton (and an interaction similar to a singleton) 
and to move from using specific classes to using interfaces.

There are two interfaces introduced: MessageSender and MessageMetadataRepository.  A class that implements the 
MessageMetadataRepository interface allows one to store simple metadata against an outgoing message (for example, the 
duration for which resulting metrics should last) and later query that metadata when the return message is received.

The MessageSender interface describes the ability to send an outgoing message and store certain metadata for any return 
message.  It is anticipated that a MessageSender will have access to a MessageMetadataRepository.

A large part of this patch is adjusting constructor of data-gathering classes so they can operate without referring to the 
singleton InfoProvider class (when sending a message) or to obtain the MessageHandlerChain (when querying message-related 
metadata).

Acked-by: Gerd
RB: http://rb.dcache.org/r/661/