The StateChangeSet object contains information about what is going to change in a given StateComposite (i.e., branch).
We'll need to add support for purging everything within a subtree and then add new metrics.
To do this, we'll need to mark children as "to be removed" then later change our minds for
some subset of them and mark these StateComponents as accept a new value (an update).  This
is in contrast with current StateChangeSet behaviour: StateChangeSet currently assumed that
a StateComponent would be assigned only a single operation (creating a new StateComponent,
removing a StateComponent, updating a StateComponent).

With this patch, a StateComponent may be marked for any operation then this changed later on.

The patch also includes a new method ensureChildNotRemoved() that prevents a child from being
removed.

Acked-by: Gerd
RB: http://rb.dcache.org/r/670/