Minor rewrite: change expiry behaviour when date expires
This patch alters how StateComposite (i.e. branches) builds up a StateTransition when metrics or
branches are to expire.

The majority of the code change is trivial, just making it easier to read.

There is a change in behaviour when the date exactly matches the expiry date.  Previously the metric
or branch wouldn't be expired.  With the patch the metric/branch is removed.  This eliminates a
race condition and aids unit testing.

Acked-by: Gerd
RB: http://rb.dcache.org/r/669/