Add LinkInfoVisitor class to extract LinkInfo information
This patch adds the LinkInfoVisitor class.  This visitor allows scanning dCache
state to obtain a mapping between a link's name and a corresponding LinkInfo
object.

Since it is anticipated that this class will mostly be used to query current or
future (subsequent to applying a StateTransition) dCache state a couple of static
methods are provided that create a new LinkInfoVisitor, query the provided
StateExhibitor and return the resulting mapping.

Acked-by: Gerd
RB: http://rb.dcache.org/r/668