pool: Add some protection against overcommitting pools
With Java 6 we can query the amount of free space of a file system. This
patch extends the periodic health check of a pool such that the amount
of expected free space is compared to the actual amount of free space.
If the latter is smaller than the former, then the pool size is adjusted.

As the health check now does more, I have increased the period of the
check from 30 seconds to 60 seconds.

Notice that the patch doesn't prevent the pool from running out of
space. Since the check only runs once a minute, there is still a risk
that the pool runs out of space before the health check can adjust the
amount of available space. In addition, the health check doesn't know
how much space was already "checked out" by a mover but not yet consumed
on the file system. Hence it may seem like there is enough space, but
really we already reserved some of that space for a particular file.
Luckikly the gap parameter reduces the risk of the pool from filling up
completely.

It would be nice to prevent the administrator from choosing a too
large pool size in the first place. Unfortunately in the current
architecture, this is not possible: The Account object is unaware of the
physical file system and thus cannot check how large it is.

It becomes quite evident that MetaDataStore is misnamed. As it nowadays
also encapsulates the FileStore, it should be called something like
ReplicaStore and MetaDataRecord should be called Replica. I will rename
those classes in a follow up patch.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/663/