Add support in StateComponent to support purge.
This patch adds the buildPurgeTransition() method to StateComponents and (therefore) also all 
classes that implement this interface.  The purge transition is used to register that some 
StateComposite (branch) and all child entries are to be removed.

A purge is not indelible.  Subsequent registering of new metric values is allowed.  The resulting 
StateChangeSet will record that an updated value is to be included.

Acked-by: Gerd
RB: http://rb.dcache.org/r/672/