pool: Fix race condition in migration module
The state machine of a migration module Task used "success", "noroute",
"timeout" and "failure" as event names for most operations. In some
cases it was possible that the the events of one operation could be
interpreted as a result of another operation. To avoid this problem,
this patch prefixes these events with the operation name.

To implement this change, the Callback class has to make use of
reflection to call the correct event method.

Target: trunk
Target: 1.9.4
Target: 1.9.3
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/675/