Add ReservationInfoVisitor class for extracting information about SRM reservations
This patch introduces the ReservationInfoVisitor class.  This class implements the
StateVisitor interface so it can discover information about dCache.  The class builds
up a list of currently known SRM reservations that it provides as a mapping between
the reservation ID and a corresponding ReservationInfo object.

A couple of static methods are included to facilitate the querying of dCache state.

Acked-by: Gerd
RB: http://rb.dcache.org/r/673/