Add StateLocation class
The StateLocation object is a utility class.  It understands dCache's location of
metrics and provides static methods for manipulating test objects.  For example,
it can update a TestStateExhibitor so it records a named pool with certain metrics.
It can also update MalleableStateTransition objects to simulate the addition, update
or removal of metrics.

It's main use is to simply writing unit tests by abstracting out the location
information needed when simulating transitions that happen on a real dCache
info service.

Acked-by: Gerd
RB: http://rb.dcache.org/r/682/