srm: Reimplements login broker query
This patch replaces the current code for performing periodic queries to
the LoginBroker with a simpler version based on CellStub.

The old code used a CellMessageAnswerable callback, however the code was
not really asynchronous: The thread that send the message would
immediately block until it got a signal from the callback that the
answer had arrived.

The new version makes this explicit by using a blocking sendAndWait. The
behaviour is exactly the same as before, but the new code is much
simpler.

The patch replaces related collections with type safe generic versions.

Testing Done:

This patch was included in the upgrade we deployed at NDGF on the 8th of
September and has been running in production since.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/690/