Add NormalisedAccessSpaceMaintainer class
This patch adds the ability to calculate NAS metrics for a dCache instance using the NormalisedAccessSpaceMaintainer class: a StateWatcher that provides secondary
information (so it is a SIP).  From the JavaDoc:

A NAS is a set of pools.  They are somewhat similar to a poolgroup except that
dCache PSU has no knowledge of them. Each NAS has the following properties:

    all pools are a member of precisely one NAS,

    all NAS have at least one member pool,

    all pools within a NAS have the same "access", where access is the set
        of units that might select these pools and the set of operations permitted
        through the link.

Because of the way NAS are constructed, all pools of a single NAS have
well-defined relationship to the links: for each link, either all pools of
the same NAS are accessible or all pools are not accessible.

NAS association with a link is no exclusive. Pools in a NAS that is
accessible from one link may be accessible from a different link. Those
pools accessible in a NAS accessible through a link may share that link
with other pools.

Signed-off by: Gerd
RB: http://rb.dcache.org/r/689/