xrootd: Fixed timeout handling and cleanup for write
This patch fixes what happens when write failures are detected by the
xrootd door. Previously a newly created file would be left behind if
PoolManager selection or the pool triggered an error. This patch updates
the XrootdRedirectHandler such that in case of writes, errors will cause
the name space entry to be removed again.

The patch also solves a timeout issue wrt. writing: Due to the
aggressive timeout in Xrootd clients (10 seconds), the timeouts in the
door for submitting a mover to the pool was 3 seconds. In case of
failure, this would give us time to retry another pool. However for
write, this would typically cause two write mover to be created. Hence
this patch disables the retry on timeout for writes. The patch also
relaxes the timeout to 4 seconds to reduce the risk of triggering it. 4
seconds still gives us time to try another pool in case of errors.

Notice that once the client is successfully redirected to a pool, the
door is out of the game: Any failure after this point will not cause the
file to be deleted.

As the old behaviour would leave behind name space entries and broken
files on pools, I consider this patch a bug fix and submit it for
inclusion in 1.9.4.

Target: trunk
Target: 1.9.4
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/709/