xrootd: New xrootd data server for the pool
This patch replaces the xrootd mover with a Netty based implementation.
Due to the use of Netty, all socket I/O is asynchronous. Disk I/O is
blocking, as that is a limitation of NIO. The same file can be opened
multiple times. For each open file a FileDescriptor instance is
maintained.

There are still plenty of things to do, but the current version is
functional. The old mover still exists and can be activated by adjusting
the movermap.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/391/