srm: Fix srmMv for the case with identical source and destination
The SRM 2.2 spec says:

d) Moving an SURL to itself results in no operation and SRM_SUCCESS must
be returned for no operation.

In revision r12237 this broke when srmMv was reimplemented. The error is
that we test for equality of the two paths after we test for the
existence of the destination.

The fix is to move the test. At the same time I removed the log call, as
there is nothing special about moving a file to itself, so there is no
reason to log it.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/710/