dcap - This patch is a repackaging of dcachedcap, the changes are related to two files the spec file and the gsitunnels main Makefile.
This patch is a repackaging of dcachedcap, the changes are related to two files the spec file and the gsitunnels main Makefile.

Renamed the devel package to libdcachedcap-devel as this is better consistent with libdcachedcap.
Added libdcachedcapgsi package for the gsi plugin.
Added "make clean" to the building process in case the globing of the source included build artifacts.
Added the concepts of ARCH and FLAVOR to the linking process.
Added the targets dist rpm install bindist to the Makefile for the gsi tunnel.

ARCH is required to differentiate between 64 bit and 32 bit architectures of globus libraries.
FLAVOR is required to make things compatible to EPEL packaged unflavored globus.


Signed-off by: Owen
http://rb.dcache.org/r/703/