Allow access to the info service XML Conduit from only the loopback device
The XML conduit provides easy access to dCache state.  The XML serialiser is
used to dump the entire dCache state when a TCP socket is connected to port 22112.
This was intended as an example conduit and to provide a mechanism to allow access
to dCache state independent to the web interface.

This patch is for all supported branches: 1.9.1, 1.9.2, 1.9.3, 1.9.4 and trunk.

Acked-by: Tigran
RB: http://rb.dcache.org/r/718/