Fix race condition between sending request for data and adding metadata for resulting metrics
The data-gathering scheduler sends requests for fresh data.  When doing this it records some metadata for any 
resulting metrics.  Currently this metadata is limited to just the time-to-live for those metrics; i.e, how 
many seconds the metrics should survive in the dCache state tree before being garbage collected.

The current code has a race-condition.  The message requesting fresh data is sent before the metric metadata 
is recorded.  This creates a time window where the message thread attempts to translate an incoming message 
to a StateUpdate object (that requires the metric TTL value) whilst the data-gathering scheduler still hasn't 
added the metric to the MessageMetadataRepository.

The patch fixes this by storing the metadata first then sending the message.

Acked-by: Tigran
RB: http://rb.dcache.org/r/725/