Use correct message OUID when locating metric metadata
This patch fixes a bug where the wrong OUID is used to retrieve stored metadata for the resulting metrics.

The patch also moves where the point where the metadata entry is removed; this is to reduce the likelihood of 
stale metadata entries that must then be cleaned up.

Acked-by: Tigran
RB: http://rb.dcache.org/r/724/