slf4j: Move slf4j packages out of opensaml and upgrade to version 1.5.8
The Simple Logging Facade is a logging abstraction layer used by some
libraries. This patch moves the slf4j jars out of the opensaml package
and upgrades it to version 1.5.8. We replace the simple slf4j module
(log to stderr) with the log4j package (log to log4j).

This will be needed for use with Jetty.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/731/