PnfsManager: Send AddCacheLocation notification on PnfsSetFileAttributes
PnfsSetFileAttributes makes it possible to add cache locations without
sending PnfsAddCacheLocation. ReplicaManager relies on a
PnfsAddCacheLocation notification when cache locations are added.

This patch extends the processing of PnfsSetFileAttributes processing
such that a PnfsAddCacheLocation notification is generated for each
cache location that is added. This should make ReplicaManager functional
in current trunk.

This patch is an alternative to patch 723.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/743/