Remove unnecessary exception thrown under unusual circumstances
The predicateHasBeenTriggered() is a recursive query descending into the dCache state.  It 
returns true if the StateTransition results in an element at the specified point has changed or 
false otherwise.

The current implementation throws an exception if the path is "impossible".  For example, if 
there is a new, updated or removed metric with path aa.bb.cc and the search is for aa.bb.cc.dd 
then, since the metric "cc" cannot have children, the search path is invalid and an exception 
is thrown.

There is a problem if a non-end-element wildcard of a predicate searches a branch that contains 
both metrics and branches; for example, suppose a transition adds two metrics: aa.bb.metric1 
and aa.bb.cc.metric2.  A search for aa.bb.*.metric2 will try aa.bb.metric1 branch and so throw 
an exception, which isn't the right result.

For this reason, it is better to define that "too long" do not result in a match but also are 
not an error.  With the initial example, an added metric aa.bb.cc will not trigger the 
predicate aa.bb.cc.dd.

With this definition, nothing will throw the MetricStatePathException, so removing this from 
the interface makes sense.

Acked-by: Tigran
RB: http://rb.dcache.org/r/753/