pool: Remove task ID from migration module
Revision r11848 added an UUID to all migration module messages. That
UUID was to replace task ID for identifying migration tasks. This patch
removes the task ID. The patch breaks compatibility with pools older
than 1.9.4.

The Task class still has an ID field which is used as part of the user
interface. This field will not be removed.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/659/