Lionel Patch: to have the logs and pid files for Chimera at the same place as for other dcache cells
Lionel found useful to have the logs and pid files for Chimera at the same place as for other dcache cells.
Please find attached a patch for the startup file.
Tested against 1.9.4-3

I cleaned up the patch a little bit to get it through RB, sensible simple
cahnges to the original.

http://www.dcache.org/rt/Ticket/Display.html?id=5118
http://rb.dcache.org/r/777/
Acked-by: Tigran