For http://rb.dcache.org/r/790.
There are places in srm code that cannot handle the null authorization that would be found in the AuthorizationRecord in such a case. It is not a problem when the unauthorized role is not the first one, because srm only considers the first role. This fix prevents any null authorizations from being included in the AuthorizationRecord.