scripts: Fix generic.lib.sh script after r12375
Revision 12375 broke generic.lib.sh when $user is not defined. This is
because "$user" expands to the empty argument and thus we try to run the
program with the empty name rather than the program we actually want to
run.

Thanks to Dmitry and Tigran for reporting this.

Target: trunk
Target: 1.9.5