Update support for sending Message subclasses so the metric TTL metadata is found
There is a bug in the info system where all cellmessages (sent to request information) that have a payload
that is a subclass of Message will not work since the reply message cannot find the metric time-to-live
metadata.  The metadata is keyed against the outgoing message's UOID; however, the CellNucleus bumps the
outgoing message's _umid field and the reply CellMessage has it's _lastUmid field copied from this modified
value.

String requests (CellMessages with a String payload) are not affected by this problem because they use the
CellMessageAnswerable callback mechanism.  This provides both the outgoing and reply CellMessages.  The
outgoing CellMessage has had its _umid field bumped; however, the outgoing CellMessage's _lastUmid is not
modified and retains the original UOID value.  This is used to obtain the metric TTL metadata when processing
the reply.

This patch adds an implementation of CellMessageAnswerable in MessageHandlerChain.  If MessageHandlerChain is
asked to send a message with a String payload then it must be given a non-null CellMessageAnswerable.  If the
outgoing request's payload is a subclass of Message then the CellMessageAnswerable will be null.  With this
patch, such outgoing requests are sent with MessageHandlerChain as the CellMessageAnswerable.  The reply
messages are then received via the MessageHandlerChain's answerArrived() method, which provides sufficient
information to obtain the metric TTL metadata.

As a side-effect, this patch also tidies up the code-path for processing the replies from such requests: they
no long go through the Cell (the InfoProvider class which extends CellAdaptor).

Acked-by: Gerd
RB: http://rb.dcache.org/r/802/