Make sure we use the canonical name when registering with a LoginBroker
The previous code would use getHostName() to determine an interface's name.  This might be the interface's 
canonical name or it might not, depending on whether getCanonicalName() had been called in the past.

The patch switches this to use getCanonicalName() to ensure that the correct (canonical) name is used.

Acked-by: Gerd
RB: http://rb.dcache.org/r/812/