pool: Add checksum verification on update
Revision r12339 added the infrastructure necessary to verify checksums
when updating existing replicas with the migration module, however that
patch did not contain the options to actually enable this.

This patch adds the -verify option to the migration module. When
specified, the checksum on update of existing replicas is verified.

I submit this for inclusion in 1.9.5, as this was by some considered an
important thing to have in the 1.9.5 release.

Target: trunk
Target: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/811/