Remove hard-coded registration of the info cell as well-known.
As pointed out by Gerd, the info cell has hard-coded behaviour of registering itself as well-known.  This is bad, 
especially as there is ready support for making this an option in the existing code-base.

This small patch fixes this, making the registration process optional.  The batch file for the infoDomain is also 
updated to configure info to register itself, so this patch should have no impact on whether info is well-known by 
default.

Acked-by: Gerd
RB: http://rb.dcache.org/r/817/