srm: Fix infinite job reload loop.
This patch fixes an infinite recursion triggered by restoring expired
jobs. Before getJob would call expireRestoredJobOrCreateExpirationTimer,
which would immediately expire a Job if its lifetime was non-positive.
That in turn would trigger a state change listener to be called, which
in turn will cause getJob to be called again. That causes the infinite
recursion.


This patch changes the logic in expireRestoredJobOrCreateExpirationTimer
such that it will always schedule an expiration timer. The delay may
possibly be zero if the job has already expired and the timer will thus
immediately trigger the expiration task. The point is however that the
state change will happen on a different thread, thus breaking the
recursion.

While studying the code, I cleaned up the iteration over the job
storages such that a CopyOnWriteArrayList is used; this avoids that the
array is copied on every restore.

Target: trunk
Target: 1.9.5
Target: 1.9.4
Target: 1.9.3
Target: 1.9.2
Target: 1.9.1
Ticket: #5112
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/819/