Adjust stop and restart, and add condrestart operations to the dcache script
The stop and restart operations in the /opt/d-cache/bin/dcache script currently will attempt to stop all services 
irrespective of whether they are running or not.  If a service is not running then an error message is emitted.

This patch changes the behaviour of stop and restart so they only attempt to stop the processes that are currently 
running, irrespective of whether the service is configured to run.  The most immediate effect is that, should a 
service that is normally configured to run is no longer running (e.g., site-admin has manually stopped it) then a 
cryptic error message is no longer reported.

The patch also introduces a new command: condrestart.  This command is similar to restart in that it stops all 
running services; however, instead of starting all configured services, condrestart starts only those services that 
were running when "dcache condrestart" was running.  Assuming no errors in dCache, this will always result in the 
same number of services running after.

Besides its potential utility to the site-admin, another motivation for introducing the condrestart is that we can 
allow sites to automate the RPM upgrade procedure under certain circumstances.  However, this patch deliberately 
does not change RPM behaviour; if such a change is desirable then it must be carefully coordinated with site-admins 
as otherwise a change in the upgrade procedure will likely cause problems.

Acked-by: Gerd, Tigran
RB: http://rb.dcache.org/r/815/